Moved context definition config from Rule to ExpressionBase
