+++ src/Engine/ExpressionBase.php
@@ -7,6 +7,7 @@
 
 namespace Drupal\rules\Engine;
 
+use Drupal\Core\Plugin\Context\ContextDefinition;
 use Drupal\Core\Plugin\ContextAwarePluginBase;
 use Drupal\rules\Context\ContextProviderTrait;
 
@@ -23,6 +24,60 @@ abstract class ExpressionBase extends ContextAwarePluginBase implements Expressi
    * @var array
    */
   protected $configuration;
+
+  /**
+   * Overrides the parent constructor to populate context definitions.
+   *
+   * Expression plugins can be configured to have arbitrary context definitions.
+   *
+   * @param array $configuration
+   *   The plugin configuration, i.e. an array with configuration values keyed
+   *   by configuration option name. The special key 'context_definitions' may
+   *   be used to initialize the context definitions by setting it to an array
+   *   of definitions keyed by context names.
+   * @param string $plugin_id
+   *   The plugin_id for the plugin instance.
+   * @param mixed $plugin_definition
+   *   The plugin implementation definition.
+   */
+  public function __construct(array $configuration, $plugin_id, $plugin_definition) {
+    if (isset($configuration['context_definitions'])) {
+      $plugin_definition['context'] = $this->createContextDefinitions($configuration['context_definitions']);
+    }
+    parent::__construct($configuration, $plugin_id, $plugin_definition);
+  }
+
+  /* Converts a context definition configuration array into an object.
+   *
+   * @todo This should be replaced by some convenience method on the
+   *   ContextDefinition class in core?
+   *
+   * @param array $configuration
+   *   The configuration properties for populating the context definition
+   *   object.
+   *
+   * @return \Drupal\Core\Plugin\Context\ContextDefinitionInterface[]
+   *   A list of context definitions keyed by the context name.
+   */
+  protected function createContextDefinitions(array $configuration) {
+    $context_definitions = [];
+    foreach ($configuration as $context_name => $definition_array) {
+      $definition_array += [
+        'type' => 'any',
+        'label' => NULL,
+        'required' => TRUE,
+        'multiple' => FALSE,
+        'description' => NULL,
+      ];
+
+      $context_definitions[$context_name] = new ContextDefinition(
+        $definition_array['type'], $definition_array['label'],
+        $definition_array['required'], $definition_array['multiple'],
+        $definition_array['description']
+      );
+    }
+    return $context_definitions;
+  }
 
   /**
    * Executes a rules expression.
